fixing socket io usage and remvoing extra prints
