+++ src/backend/input/consumer.py
@@ -53,7 +53,6 @@ def process_data_live() -> None:
 
     global start_consume_time, last_consume_time
     start_consume_time = time.perf_counter()
-    print(f"Starting live data processing at time: {start_consume_time}")
     db_conn = DbConnection()
         
     # For tracking processed messages
@@ -76,7 +75,6 @@ def process_data_live() -> None:
                 # Log status every 1000 messages or 10 seconds
                 current_time = time.perf_counter()
                 if message_count % 1000 == 0 or (current_time - last_status_time) > 10:
-                    print(f"Processed {message_count} messages. Last timestamp: {can_msg.timeStamp:.3f}s")
                     last_status_time = current_time
 
         if list_can_messages:
