+++ src/main.py
@@ -11,8 +11,8 @@ from backend.db_connection import DbConnection
 from backend.input import consumer, logfile_producer, live_log_producer, radio_producer
 from functools import partial
 
-# must import sockio/ files to register the socketio event handlers
-from backend.sockio import debug_dashboard, alert_manager  
+# must import sockio/ files to register the socketio event handlers and routes
+from backend.sockio import debug_dashboard, alert_manager, routes, handlers  
 
 
 def main():
